fix stun utils
