+++ iroh-relay/src/protos/stun.rs
@@ -148,137 +148,10 @@ pub fn parse_response(b: &[u8]) -> Result<(TransactionId, SocketAddr), Error> {
 }
 
 #[cfg(test)]
-pub(crate) mod tests {
-    use std::{
-        net::{IpAddr, Ipv4Addr},
-        sync::Arc,
-    };
-
-    use anyhow::Result;
-    use tokio::{
-        net,
-        sync::{oneshot, Mutex},
-    };
-    use tracing::{debug, trace};
+mod tests {
 
     use super::*;
-    use crate::{RelayMap, RelayNode, RelayUrl};
-
-    // TODO: make all this private
-
-    /// A drop guard to clean up test infrastructure.
-    ///
-    /// After dropping the test infrastructure will asynchronously shutdown and release its
-    /// resources.
-    // Nightly sees the sender as dead code currently, but we only rely on Drop of the
-    // sender.
-    // TODO(@divma): is this just a oneshot??
-    #[derive(Debug)]
-    #[allow(dead_code)]
-    pub struct CleanupDropGuard(pub(crate) oneshot::Sender<()>);
-
-    // (read_ipv4, read_ipv5)
-    #[derive(Debug, Default, Clone)]
-    pub struct StunStats(Arc<Mutex<(usize, usize)>>);
-
-    impl StunStats {
-        pub async fn total(&self) -> usize {
-            let s = self.0.lock().await;
-            s.0 + s.1
-        }
-    }
-
-    pub fn relay_map_of(stun: impl Iterator<Item = SocketAddr>) -> RelayMap {
-        relay_map_of_opts(stun.map(|addr| (addr, true)))
-    }
-
-    pub fn relay_map_of_opts(stun: impl Iterator<Item = (SocketAddr, bool)>) -> RelayMap {
-        let nodes = stun.map(|(addr, stun_only)| {
-            let host = addr.ip();
-            let port = addr.port();
-
-            let url: RelayUrl = format!("http://{host}:{port}").parse().unwrap();
-            RelayNode {
-                url,
-                stun_port: port,
-                stun_only,
-            }
-        });
-        RelayMap::from_nodes(nodes).expect("generated invalid nodes")
-    }
-
-    /// Sets up a simple STUN server binding to `0.0.0.0:0`.
-    ///
-    /// See [`serve`] for more details.
-    pub(crate) async fn serve_v4() -> Result<(SocketAddr, StunStats, CleanupDropGuard)> {
-        serve(std::net::Ipv4Addr::UNSPECIFIED.into()).await
-    }
-
-    /// Sets up a simple STUN server.
-    pub(crate) async fn serve(ip: IpAddr) -> Result<(SocketAddr, StunStats, CleanupDropGuard)> {
-        let stats = StunStats::default();
-
-        let pc = net::UdpSocket::bind((ip, 0)).await?;
-        let mut addr = pc.local_addr()?;
-        match addr.ip() {
-            IpAddr::V4(ip) => {
-                if ip.octets() == [0, 0, 0, 0] {
-                    addr.set_ip("127.0.0.1".parse().unwrap());
-                }
-            }
-            _ => unreachable!("using ipv4"),
-        }
-
-        println!("STUN listening on {}", addr);
-        let (s, r) = oneshot::channel();
-        let stats_c = stats.clone();
-        tokio::task::spawn(async move {
-            run_stun(pc, stats_c, r).await;
-        });
-
-        Ok((addr, stats, CleanupDropGuard(s)))
-    }
-
-    async fn run_stun(pc: net::UdpSocket, stats: StunStats, mut done: oneshot::Receiver<()>) {
-        let mut buf = vec![0u8; 64 << 10];
-        loop {
-            trace!("read loop");
-            tokio::select! {
-                _ = &mut done => {
-                    debug!("shutting down");
-                    break;
-                }
-                res = pc.recv_from(&mut buf) => match res {
-                    Ok((n, addr)) => {
-                        trace!("read packet {}bytes from {}", n, addr);
-                        let pkt = &buf[..n];
-                        if !is(pkt) {
-                            debug!("received non STUN pkt");
-                            continue;
-                        }
-                        if let Ok(txid) = parse_binding_request(pkt) {
-                            debug!("received binding request");
-                            let mut s = stats.0.lock().await;
-                            if addr.is_ipv4() {
-                                s.0 += 1;
-                            } else {
-                                s.1 += 1;
-                            }
-                            drop(s);
-
-                            let res = response(txid, addr);
-                            if let Err(err) = pc.send_to(&res, addr).await {
-                                eprintln!("STUN server write failed: {:?}", err);
-                            }
-                        }
-                    }
-                    Err(err) => {
-                        eprintln!("failed to read: {:?}", err);
-                    }
-                }
-            }
-        }
-    }
+    use std::net::{IpAddr, Ipv4Addr};
 
     // TODO(@divma): what's this and why is it commented?
     // Test to check if an existing stun server works
